reorganize new code
