+++ backend/src/main.py
@@ -5,6 +5,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from .db import DBWrapper
 from .handlers import ada, alert, ridership, routes, stops, vans
 from .hardware import HardwareExceptionMiddleware
+from .new.routes.alerts import router
 from .vantracking.factory import van_tracker
 from .vantracking.tracker import VanTracker
 
@@ -24,7 +25,8 @@ app.add_middleware(HardwareExceptionMiddleware)
 app.include_router(ada.router)
 app.include_router(routes.router)
 app.include_router(stops.router)
-app.include_router(alert.router)
+# app.include_router(alert.router)
+app.include_router(router)
 app.include_router(ridership.router)
 app.include_router(vans.router)
